+++ lib/need_for_type/file_manager.rb
@@ -1,17 +1,15 @@
 module NeedForType
   class FileManager
-    attr_accessor :content
+    attr_accessor :text, :text_id
 
     def initialize()
       @path = "#{NeedForType::GEM_ROOT}/assets/texts/"
     end
 
-    # Reads random file from difficulty folder
-    def get_random_text
+    def read_random_text
       directory_count = Dir[File.join(@path, '**', '*')].count
-      random_file = rand(1..directory_count)
-
-      @content = File.read("#{@path}/text-#{random_file}.txt").strip
+      @text_id = rand(1..directory_count)
+      @text = File.read("#{@path}/text-#{@text_id}.txt").strip
     end
   end
 end