+++ lib/need_for_type/states.rb
@@ -3,6 +3,7 @@ require 'need_for_type/states/menu'
 require 'need_for_type/states/game'
 require 'need_for_type/states/start'
 require 'need_for_type/states/end'
+require 'need_for_type/states/submit_score'
 
 module NeedForType
   module States